removed storing and publishing of the bounding box history
For every hypothesis, the bounding box in every time step was saved to publish those to a filter package.
The filter package would receive the boxes and remove every point of a point cloud that lays within one of theses boxes.
This was used to filter points corresponding to dynamic objects from a point cloud map.