Removed functionality to track humans
This use case was very specific and led to a dependency on the person_msgs package from another repository.
Removing the dependency and functionality cleans up this package, which is supposed to be a generic package for multi hypothesis tracking.
The removed functionality can be found in the multi_human_tracking branch.